+++ Tags/Blamtoozle/halo2_tools_pc64/halo2_tools_pc64_field_type.h
@@ -1,6 +1,6 @@
 #pragma once
 
-enum e_halo2_tools_pc64_field : unsigned long
+enum e_halo2_tools_pc64_field : unsigned int
 {
 	_halo2_tools_pc64_field_string,
 	_halo2_tools_pc64_field_long_string,
