+++ Tags/Blamtoozle/halo4_x360/halo4_x360_tag_struct_definition.h
@@ -28,7 +28,7 @@ enum e_halo4_x360_tag_memory_usage_bit
 //	return (this->runtime_flags >> 6) & 1;
 //}
 
-enum e_halo4_x360_tag_field_set_bit : unsigned long
+enum e_halo4_x360_tag_field_set_bit : unsigned int
 {
 	_halo4_x360_tag_field_set_unknown0,
 	_halo4_x360_tag_field_set_unknown1,
